+++ packages/@expo/metro-config/src/serializer/getCssDeps.ts
@@ -75,6 +75,7 @@ export function getCssSerialAssets<T extends any>(
         for (const external of cssMetadata.externalImports) {
           let source = `<link rel="stylesheet" href="${escapeHtmlAttribute(external.url)}"`;
 
+          // TODO(@hassankhan): We should be able to remove this when we remove the static renderer
           // TODO: How can we do this for local css imports?
           if (external.media) {
             source += ` media="${escapeHtmlAttribute(external.media)}"`;
@@ -92,6 +93,9 @@ export function getCssSerialAssets<T extends any>(
             source,
             metadata: {
               hmrId: pathToHtmlSafeName(originFilename),
+              // Carried alongside the baked `source` field so the streaming renderer can rebuild
+              // the `<link>` as a React node without having to parse the HTML string
+              media: external.media ?? undefined,
             },
           });
         }